refactor of controllers in mvc, moved /relations request mapping related methods to a separate controller.
