+++ src/main/kotlin/com/ritense/iko/mvc/controller/TestController.kt
@@ -43,7 +43,7 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.servlet.ModelAndView
 import java.util.UUID
 
-@Controller
+@Controller("mvcTestController")
 @RequestMapping("/admin")
 internal class TestController(
     private val producerTemplate: ProducerTemplate,
